Revert vars section to json before parsing it by YAML.
For this we use serde_json::Map instead of HashMap.
serde_json::Map is one of possible variant of serde_json::Value.
So we only need to cast it backs all variable block to this variant and pass it to the convert_json_vars.
Plus there is no need to check var by var, we can treat all the block.